> flatwidget for consumption-widget
