+++ ui/src/app/edge/live/common/consumption/consumption.component.html
@@ -1,0 +1,34 @@
+<ng-container *ngIf="isInitialized && (consumptionMeterComponents || evcsComponents)">
+    <oe-flat-widget img="assets/img/consumption.png" [title]="'General.consumption' | translate">
+        <ng-container *ngIf="evcsComponents.length > 0">
+            <ng-container *ngFor="let component of evcsComponents">
+                <ng-container *ngIf="evcsChargePower[component.id] > 0">
+                    <oe-flat-widget-line [name]="component.alias" [channelAddress]="component.id + '/ChargePower'">
+                    </oe-flat-widget-line>
+                </ng-container>
+            </ng-container>
+        </ng-container>
+        <ng-container *ngIf="evcsComponents.length == 0 && consumptionMeterComponents.length == 0">
+            <oe-flat-widget-line [value]="sumActivePower" [converter]="this.utils.CONVERT_WATT_TO_KILOWATT">
+            </oe-flat-widget-line>
+        </ng-container>
+        <ng-container *ngIf="consumptionMeterComponents.length > 0">
+            <ng-container *ngFor="let component of consumptionMeterComponents">
+                <oe-flat-widget-line [name]="component.alias" [channelAddress]="component.id + '/ActivePower'">
+                </oe-flat-widget-line>
+            </ng-container>
+        </ng-container>
+        <ng-container *ngIf="evcsComponents.length != 0 || consumptionMeterComponents.length != 0">
+            <ng-container *ngIf="otherPower <= 0, else showOtherPower">
+                <oe-flat-widget-line [name]="'General.otherConsumption' | translate" [value]="0"
+                    [converter]="this.utils.CONVERT_WATT_TO_KILOWATT">
+                </oe-flat-widget-line>
+            </ng-container>
+            <ng-template #showOtherPower>
+                <oe-flat-widget-line [name]="'General.otherConsumption' | translate" [value]="otherPower"
+                    [converter]="this.utils.CONVERT_WATT_TO_KILOWATT">
+                </oe-flat-widget-line>
+            </ng-template>
+        </ng-container>
+    </oe-flat-widget>
+</ng-container>
