+++ ui/src/app/edge/live/common/consumption/modal/modal.component.html
