+++ ui/src/app/edge/live/common/consumption/modal/modal.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { Edge, Service, EdgeConfig } from '../../../../shared/shared';
+import { Edge, Service, EdgeConfig } from '../../../../../shared/shared';
 import { ModalController } from '@ionic/angular';
 
 @Component({
