--- ui/src/app/edge/live/consumption/consumption.component.html
@@ -1,70 +1,0 @@
-<ng-container *ngIf="service.currentEdge | async as edge">
-    <ng-container *ngIf="edge.currentData | async as currentData">
-        <ion-card button (click)="presentModal()">
-            <ion-item lines="full" color="light">
-                <ion-avatar slot="start">
-                    <img src="assets/img/consumption.png">
-                </ion-avatar>
-                <ion-label translate>General.consumption</ion-label>
-            </ion-item>
-            <ion-card-content *ngIf="currentData.summary.consumption as sum">
-                <table class="full_width" *ngIf="evcsComponents">
-                    <ng-container *ngIf="evcsComponents.length == 0 && consumptionMeterComponents.length == 0">
-                        <tr>
-                            <td style="width:65%"></td>
-                            <td style="width:35%" class="align_right">
-                                {{ sum.activePower | unitvalue:'kW' }}
-                            </td>
-                        </tr>
-                    </ng-container>
-                    <ng-container *ngIf="(edge.currentData | async)['channel'] as currentData">
-                        <ng-container *ngIf="evcsComponents.length > 0">
-                            <ng-container *ngFor="let component of evcsComponents">
-                                <ng-container *ngIf="currentData[component.id + '/ChargePower'] > 0">
-                                    <tr>
-                                        <td style="width:65%" *ngIf="component.id == component.alias">{{component.id}}
-                                        </td>
-                                        <td style="width: 65%" *ngIf="component.id != component.alias">
-                                            {{component.alias}}
-                                        </td>
-                                        <td style="width:35%" class="align_right">
-                                            {{ currentData[component.id + '/ChargePower'] | unitvalue:'kW' }}
-                                        </td>
-                                    </tr>
-                                </ng-container>
-                            </ng-container>
-                        </ng-container>
-                        <ng-container *ngIf="consumptionMeterComponents.length > 0">
-                            <ng-container *ngFor="let component of consumptionMeterComponents">
-                                <ng-container *ngIf="currentData[component.id + '/ActivePower'] > 0">
-                                    <tr>
-                                        <td style="width:65%" *ngIf="component.id == component.alias">{{component.id}}
-                                        </td>
-                                        <td style="width: 65%" *ngIf="component.id != component.alias">
-                                            {{component.alias}}
-                                        </td>
-                                        <td style="width:35%" class="align_right">
-                                            {{ currentData[component.id + '/ActivePower'] | unitvalue:'kW' }}
-                                        </td>
-                                    </tr>
-                                </ng-container>
-                            </ng-container>
-                        </ng-container>
-                    </ng-container>
-
-                    <tr *ngIf="evcsComponents.length != 0 || consumptionMeterComponents.length != 0">
-                        <td style="width: 65%" translate>General.otherConsumption</td>
-                        <ng-container *ngIf="sum.activePower - getTotalOtherPower() as otherPower">
-                            <td style="width: 35%" class="align_right" *ngIf="otherPower > 0">
-                                {{ otherPower | unitvalue:'kW' }}
-                            </td>
-                            <td style="width: 35%" class="align_right" *ngIf="otherPower <= 0">
-                                {{ 0 | unitvalue:'kW' }}
-                            </td>
-                        </ng-container>
-                    </tr>
-                </table>
-            </ion-card-content>
-        </ion-card>
-    </ng-container>
-</ng-container>
